requirements.txt
lxml==4.6.3
openpyx==2.4.11
Evtx==0.7.3
